Fix relocations, Add data relocations and change section names
objdump likes having a null value first, apparently. not sure why,
but i will do it for you, objdump - it's only 1 byte :)

relocations for data section now implemented, too! (and the data
section itself)

finally, the relocations have been fixed to be in charge of
absolute addresses instead of the main program needing to know
0x400000.